Rename private files in optuna/integration/lightgbm_tuner.
