+++ tests/integration_tests/lightgbm_tuner_tests/test_optimize.py
@@ -13,11 +13,11 @@ import pytest
 
 import optuna
 import optuna.integration.lightgbm as lgb
-from optuna.integration.lightgbm_tuner.optimize import BaseTuner
-from optuna.integration.lightgbm_tuner.optimize import LightGBMTuner
-from optuna.integration.lightgbm_tuner.optimize import LightGBMTunerCV
-from optuna.integration.lightgbm_tuner.optimize import OptunaObjective
-from optuna.integration.lightgbm_tuner.optimize import OptunaObjectiveCV
+from optuna.integration.lightgbm_tuner._optimize import BaseTuner
+from optuna.integration.lightgbm_tuner._optimize import LightGBMTuner
+from optuna.integration.lightgbm_tuner._optimize import LightGBMTunerCV
+from optuna.integration.lightgbm_tuner._optimize import OptunaObjective
+from optuna.integration.lightgbm_tuner._optimize import OptunaObjectiveCV
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
@@ -701,7 +701,7 @@ class TestLightGBMTuner(object):
         def objective(trial: optuna.trial.Trial, value: float) -> float:
 
             trial.set_system_attr(
-                optuna.integration.lightgbm_tuner.optimize._STEP_NAME_KEY,
+                optuna.integration.lightgbm_tuner._optimize._STEP_NAME_KEY,
                 "step{:.0f}".format(value),
             )
             return trial.suggest_uniform("x", value, value)
